Hopefully starting to see transactions soon.
Some magical stuff has to happen first. We generate bitcoins in bitcoind - which
end up being part of whatever default account it provisions itself with.

Then we need to use raw transactions to generate the transfer - otherwise it
assuems the source of the money is bitcoind wallet - which we should never ever
use for our users.