Repeatable test bench setup.
The performance here can be dramatically improved by simply restore a known
~/.bitcoin/regtest folder and restarting between runs.

Some useful bits and pieces still come out of doing it this way. Namely the
discovery that bitcoind's built in private key is not interchangeable with and
HD key. That's documented but it's good to see it fail :)

The code is still very raw and resembles a shell script rather than the library
it should grow into. Most of what lies ahead of us at this point will be managing
public and private keys. With HD keys we can derived multiple addresses which
will necessitate a different approach to things like "listunspent" that can take
an array of addresses - how well does this scale after a users has done hundreds
of transfers and has potentially hundreds of corresponding addresses?

Responding to feedback.