About page.
A few minor changes here. I added a gulpfile back just so I can have it watching
my scss file for changes. I think one of the weaknesses of duo as a system for
sass is that it's quite slow to render and audition changes - a typical
gulp-watch task would be clever enough to only compile a single file.

I also broke out the vendor css into it's own file. With @import we're rolling it
all into a single file multiple times for each of app / marketting.