Some bug fixes and random stuff.
Moved a few pieces out of the kernel and into the server. It wanted to load
the kernel to have you logging + config environment and it resulted in a circular
import.

When looking into the e-mail signup I wanted to log some form output. The logger
is cool but it was filled with self calls from goconvoy (which appears to be
doing magical things!)